chore: move transformPromptWithContext to separate file
