+++ src/utils/llmPrompt.ts
@@ -1,0 +1,43 @@
+import { getPref } from "./prefs";
+import { TranslateTask } from "./task";
+
+export function transformPromptWithContext(
+  prefKey: string,
+  langFrom: string,
+  langTo: string,
+  sourceText: string,
+  data: Required<TranslateTask>,
+): string {
+  let prompt = getPref(prefKey) as string;
+
+  if (getPref("attachPaperContext") && data.itemId) {
+    const item = Zotero.Items.get(data.itemId);
+    const topItem = item ? Zotero.Items.getTopLevel([item])[0] : null;
+    if (topItem) {
+      let contextInfo = "";
+      const title = topItem.getField("title") as string;
+      const abstract = topItem.getField("abstractNote") as string;
+
+      if (title) {
+        contextInfo += `Paper Title: ${title}`;
+      }
+      if (abstract) {
+        contextInfo += title
+          ? `\n\nPaper Abstract: ${abstract}`
+          : `Paper Abstract: ${abstract}`;
+      }
+
+      if (contextInfo) {
+        prompt = prompt.replace(
+          "${sourceText}",
+          `Context from the academic paper:\n${contextInfo}\n\nText to translate: ${sourceText}`,
+        );
+      }
+    }
+  }
+
+  return prompt
+    .replaceAll("${langFrom}", langFrom)
+    .replaceAll("${langTo}", langTo)
+    .replaceAll("${sourceText}", sourceText);
+}
